+++ services/web/server/src/simcore_service_webserver/projects/_metadata_handlers.py
@@ -28,7 +28,7 @@ from ..login.decorators import login_required
 from ..security.decorators import permission_required
 from ..utils_aiohttp import envelope_json_response
 from . import _metadata_api
-from ._common.exception_handlers import handle_plugin_requests_exceptions
+from ._common.exceptions_handlers import handle_plugin_requests_exceptions
 from ._common.models import ProjectPathParams, RequestContext
 
 routes = web.RouteTableDef()
